분산 락 수정

+++ concert-service/src/main/java/com/bit/docker/concertservice/api/ConcertController.java
@@ -23,12 +23,12 @@ public class ConcertController {
     }
 
     @GetMapping("/{concertId}")
-    public Concert concert(@PathVariable Long concertId) {
+    public Concert concert(@PathVariable int concertId) {
         return concertQueryService.getConcert(concertId);
     }
 
     @GetMapping("/{concertId}/seats")
-    public List<Seat> seats(@PathVariable Long concertId) {
+    public List<Seat> seats(@PathVariable int concertId) {
         return concertQueryService.getSeats(concertId);
     }
 }
